+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [v0.9.0]
+
+- API improvements, bugfix and general improvements for CCSDS spacepacket
+  header implementation
+
 ## [v0.8.1]
 
 - Named value for fetching global APID